refactor(docs): isolate build timing hooks
