+++ src/bokeh/sphinxext/_internal/bokeh_autodoc.py
@@ -25,10 +25,9 @@ log = logging.getLogger(__name__)
 # -----------------------------------------------------------------------------
 
 # Standard library imports
-from dataclasses import dataclass, field
-from heapq import nlargest
+from dataclasses import dataclass
 from time import perf_counter
-from typing import Any, NamedTuple
+from typing import Any
 
 # External imports
 from sphinx.ext.autodoc import (
@@ -45,6 +44,7 @@ from bokeh.model import Model
 
 # Bokeh imports
 from . import PARALLEL_SAFE, SphinxParallelSpec
+from .bokeh_timing import _setup_docs_profile
 from .bokeh_toc import _shorten_reference_toc_titles
 
 # -----------------------------------------------------------------------------
@@ -128,14 +128,11 @@ def setup(app: Any) -> SphinxParallelSpec:
     app.add_autodocumenter(EnumDocumenter)
     app.add_autodocumenter(PropDocumenter)
     app.add_autodocumenter(ModelDocumenter)
-    app.connect("builder-inited", _start_docs_profile, priority=999)
-    app.connect("env-before-read-docs", _start_docs_read)
+    _setup_docs_profile(app)
     app.connect("env-updated", _shorten_reference_toc_titles)
     app.connect("env-updated", _prune_empty_viewcode_modules)
-    app.connect("env-updated", _finish_docs_read, priority=999)
-    app.connect("write-started", _start_docs_write, priority=100)
     app.connect("write-started", _cache_python_domain_fuzzy_lookups, priority=999)
-    app.connect("build-finished", _log_docs_profile, priority=999)
+    app.connect("build-finished", _log_fuzzy_lookup_stats, priority=999)
 
     return PARALLEL_SAFE
 
@@ -143,24 +140,7 @@ def setup(app: Any) -> SphinxParallelSpec:
 # Private API
 # -----------------------------------------------------------------------------
 
-_PROFILE_ATTR = "_bokeh_docs_profile"
 _FUZZY_LOOKUP_STATS_ATTR = "_bokeh_python_fuzzy_lookup_stats"
-
-
-class _DocumentTiming(NamedTuple):
-    duration: float
-    docname: str
-
-
-@dataclass
-class _DocsProfile:
-    started: float
-    read_started: float | None = None
-    read_finished: float | None = None
-    write_started: float | None = None
-    documents: int = 0
-    resolve_timings: list[_DocumentTiming] = field(default_factory=list)
-    serialize_timings: list[_DocumentTiming] = field(default_factory=list)
 
 
 @dataclass
@@ -172,94 +152,7 @@ class _FuzzyLookupStats:
     elapsed: float = 0.0
 
 
-def _start_docs_profile(app: Any) -> None:
-    setattr(app, _PROFILE_ATTR, _DocsProfile(started=perf_counter()))
-
-
-def _start_docs_read(app: Any, env: Any, docnames: list[str]) -> None:
-    profile = getattr(app, _PROFILE_ATTR, None)
-    if profile is not None:
-        profile.read_started = perf_counter()
-        profile.documents = len(docnames)
-
-
-def _finish_docs_read(app: Any, env: Any) -> None:
-    profile = getattr(app, _PROFILE_ATTR, None)
-    if profile is not None:
-        profile.read_finished = perf_counter()
-
-
-def _start_docs_write(app: Any, builder: Any) -> None:
-    profile = getattr(app, _PROFILE_ATTR, None)
-    if profile is not None:
-        profile.write_started = perf_counter()
-        profile.resolve_timings.clear()
-        profile.serialize_timings.clear()
-
-        resolve_doctree = builder.env.get_and_resolve_doctree
-
-        def timed_resolve_doctree(docname: str, *args: Any, **kwargs: Any) -> Any:
-            started = perf_counter()
-            try:
-                return resolve_doctree(docname, *args, **kwargs)
-            finally:
-                profile.resolve_timings.append(_DocumentTiming(
-                    duration=perf_counter() - started,
-                    docname=docname,
-                ))
-
-        builder.env.get_and_resolve_doctree = timed_resolve_doctree
-
-        write_serialized = builder.write_doc_serialized
-
-        def timed_write_serialized(docname: str, *args: Any, **kwargs: Any) -> Any:
-            started = perf_counter()
-            try:
-                return write_serialized(docname, *args, **kwargs)
-            finally:
-                profile.serialize_timings.append(_DocumentTiming(
-                    duration=perf_counter() - started,
-                    docname=docname,
-                ))
-
-        builder.write_doc_serialized = timed_write_serialized
-
-
-def _log_docs_profile(app: Any, exception: Exception | None) -> None:
-    finished = perf_counter()
-    profile = getattr(app, _PROFILE_ATTR, None)
-    if (
-        profile is not None
-        and profile.read_started is not None
-        and profile.read_finished is not None
-        and profile.write_started is not None
-    ):
-        read_seconds = profile.read_finished - profile.read_started
-        prepare_seconds = profile.write_started - profile.read_finished
-        write_seconds = finished - profile.write_started
-        total_seconds = finished - profile.started
-        log.info(
-            f"Bokeh Sphinx timings: read={read_seconds:.3f}s prepare={prepare_seconds:.3f}s "
-            f"write={write_seconds:.3f}s total={total_seconds:.3f}s "
-            f"documents={profile.documents} workers={app.parallel}",
-        )
-
-        resolve_timings = profile.resolve_timings
-        log.info(
-            f"Bokeh doctree resolution: documents={len(resolve_timings)} "
-            f"total={sum(timing.duration for timing in resolve_timings):.3f}s",
-        )
-        for timing in nlargest(5, resolve_timings, key=lambda timing: timing.duration):
-            log.info(f"Bokeh doctree slow: {timing.duration:.3f}s {timing.docname}")
-
-        serialize_timings = profile.serialize_timings
-        log.info(
-            f"Bokeh search-index serialization: documents={len(serialize_timings)} "
-            f"total={sum(timing.duration for timing in serialize_timings):.3f}s",
-        )
-        for timing in nlargest(5, serialize_timings, key=lambda timing: timing.duration):
-            log.info(f"Bokeh search-index slow: {timing.duration:.3f}s {timing.docname}")
-
+def _log_fuzzy_lookup_stats(app: Any, exception: Exception | None) -> None:
     stats = getattr(app.builder, _FUZZY_LOOKUP_STATS_ATTR, None)
     if stats is not None:
         reused = stats.fallbacks - stats.computed
