+++ tests/unit/bokeh/sphinxext/test_bokeh_autodoc.py
@@ -7,9 +7,7 @@ from typing import Any, cast
 # Bokeh imports
 from bokeh.sphinxext._internal.bokeh_autodoc import (
     _cache_python_domain_fuzzy_lookups,
-    _DocsProfile,
     _prune_empty_viewcode_modules,
-    _start_docs_write,
 )
 
 
@@ -81,30 +79,3 @@ def test_cache_python_domain_fuzzy_lookups() -> None:
     assert stats.fallbacks == 2
     assert stats.computed == 1
     assert stats.elapsed >= 0
-
-
-def test_start_docs_write_times_main_process_operations() -> None:
-    calls = []
-
-    def resolve(docname: str, value: int) -> tuple[str, int]:
-        calls.append(("resolve", docname, value))
-        return docname, value
-
-    def serialize(docname: str, value: int) -> tuple[str, int]:
-        calls.append(("serialize", docname, value))
-        return docname, value
-
-    env = SimpleNamespace(get_and_resolve_doctree=resolve)
-    builder = SimpleNamespace(env=env, write_doc_serialized=serialize)
-    app = SimpleNamespace(_bokeh_docs_profile=_DocsProfile(started=0.0))
-
-    _start_docs_write(cast(Any, app), cast(Any, builder))
-
-    assert env.get_and_resolve_doctree("document", 1) == ("document", 1)
-    assert builder.write_doc_serialized("document", 2) == ("document", 2)
-    assert calls == [
-        ("resolve", "document", 1),
-        ("serialize", "document", 2),
-    ]
-    assert app._bokeh_docs_profile.resolve_timings[0].docname == "document"
-    assert app._bokeh_docs_profile.serialize_timings[0].docname == "document"
